Move service to services.githubrepo
Now that there aren't additional items in services.github, it makes
sense to move the repo module into the root of the services subpackage
and rename as githubrepo to match the class name.
